Implement DGSEM on curved structured mesh in 2D (and 1D) (#498)
* Add first prototype of bilinearly mapped mesh

* Calculate correct stepsize for constant speed on curved mesh

* Implement transformed surface flux

* Port extended example to curved mesh

* Implement rotated surface flux and port euler examples

* Add examples for rotated square and parallelogram

* Fix test cases

* Add transfinite mapping for curved meshes

* Add curved examples and tests for advection and euler

* Fix rotated euler tests

* Optimize performance for curved mesh

* Make non-performance-relevant code more readable

* Remove now unneeded using StaticArrays

* Rename structured to curved mesh

* Fix mesh saving and loading for StructuredMesh

* Update news

* Fix visualization

* Fix visualization

* Adapt 1d to Curved

* Enable callbacks

* Add mesh to dispatch 1d callbacks

* Fix Formating

* Remove elements from initialize loops

* Add documentation

* Revise CurvedMesh data structure

* Save RealT for restarting

* Fix rotated flux calculation

* Remove argument RealT when coordinates are specified

* Fix stepsize for curved mesh

* Fix stepsize for CurvedMesh in 1D

* Revise initial_condition_rotated for linear advection

* Fix 3D tests

* Improve documentation of CurvedMesh

* Fix advection rotated example

* Add fancy summary box for CurvedMesh

* Improve formatting

* Improve performance of flux rotation with linear algebra skills

* Improve comments

* Implement suggestions

* Add generic rotated numerical flux

* Improve performance of calc_inverse_jacobian!

* Invert inverse jacobian before interpolation

* Remove default implementation of rotate_to/from_x

* Implement suggested changes

* Implement suggestions

+++ src/equations/compressible_euler_3d.jl
@@ -415,7 +415,7 @@ end
 
 
 # Calculate 1D flux for a single point
-@inline function flux(u, orientation, equations::CompressibleEulerEquations3D)
+@inline function flux(u, orientation::Integer, equations::CompressibleEulerEquations3D)
   rho, rho_v1, rho_v2, rho_v3, rho_e = u
   v1 = rho_v1/rho
   v2 = rho_v2/rho
