+++ src/equations/ideal_glm_mhd_3d.jl
@@ -135,7 +135,7 @@ end
 
 
 # Calculate 1D flux in for a single point
-@inline function flux(u, orientation, equations::IdealGlmMhdEquations3D)
+@inline function flux(u, orientation::Integer, equations::IdealGlmMhdEquations3D)
   rho, rho_v1, rho_v2, rho_v3, rho_e, B1, B2, B3, psi = u
   v1 = rho_v1/rho
   v2 = rho_v2/rho
