+++ src/equations/ideal_glm_mhd_multicomponent_2d.jl
@@ -159,7 +159,7 @@ end
 
 
 # Calculate 1D flux in for a single point
-@inline function flux(u, orientation, equations::IdealGlmMhdMulticomponentEquations2D)
+@inline function flux(u, orientation::Integer, equations::IdealGlmMhdMulticomponentEquations2D)
   rho_v1, rho_v2, rho_v3, rho_e, B1, B2, B3, psi = u
   @unpack c_h = equations
